Refactor bluetooth functionality
