+++ src/d39_bt/d39_bt.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void d39_bt_init(char **err);
+void d39_bt_send_msg(char *buf, size_t len, char **err);
