+++ src/d39_gpio/d39_gpio.h
@@ -1,3 +1,4 @@
+#pragma once
 
 void d39_gpio_init(char **err);
 uint32_t d39_gpio_getticks(char **err);
